+++ late-ssh/src/app/chat/mod.rs
@@ -1,5 +1,4 @@
 pub mod input;
-pub(crate) mod mentions;
 pub mod news;
 pub mod notifications;
 pub mod state;
