+++ late-ssh/src/app/common/mod.rs
@@ -1,4 +1,6 @@
 pub mod composer;
+pub mod markdown;
+pub(crate) mod mentions;
 pub mod overlay;
 pub mod primitives;
 pub mod sidebar;
